Optimization in Admin Frontend page. close #7
