+++ Admin_Teacher_Frontend/src/config.ts
@@ -1,15 +1,2 @@
-export const firebaseConfig = {
-    apiKey: "AIzaSyAjJleo9eFrhSQAERZrw0ERgGaAuGgeIIs",
-    authDomain: "dasaram-backend.firebaseapp.com",
-    projectId: "dasaram-backend",
-    storageBucket: "dasaram-backend.firebasestorage.app",
-    messagingSenderId: "535594747405",
-    appId: "1:535594747405:web:002603864da54adbbe3744",
-    measurementId: "G-QL6QM02DSS"
-};
-
 export const BACKEND_URL = 'http://localhost:3000/api/v1/free' 
-
-
-
-
+// export const BACKEND_URL = 'https://test-demo-7n6y.onrender.com/api/v1/free'
