+++ Admin_Teacher_Frontend/src/pages/Dashbord/HomePage.tsx
@@ -1,98 +1,23 @@
 "use client";
 
-import { useState } from "react";
 import { Link } from "react-router-dom";
-import { useRecoilState } from "recoil";
-import { motion, AnimatePresence } from "framer-motion";
+import {  useRecoilValue } from "recoil";
+import { motion } from "framer-motion";
 import { userAtom } from "../../state/userAtom";
-import Cookies from "js-cookie";
-import {
-  FaChalkboardTeacher,
-  FaUserGraduate,
-  FaBook,
-  FaClipboardList,
-  FaUserCheck,
-  FaBars,
-  FaSignOutAlt,
-} from "react-icons/fa";
-import { UserRole } from "../../types/type";
-import Sidebar from "./Sidebar";
 
-export const menuItems = [
-  { title: "Manage Subjects", icon: FaBook, link: "/subject-std" },
-  {
-    title: "Manage Teachers",
-    icon: FaChalkboardTeacher,
-    link: "/manage-teacher",
-  },
-  { title: "Manage Students", icon: FaUserGraduate, link: "/manage-student" },
-  { title: "Manage Tests", icon: FaClipboardList, link: "/manage-test" },
-  { title: "Attendance", icon: FaUserCheck, link: "/add-attedance" },
-];
+import { UserRole } from "../../types/type";
+import { menuItems } from "../../App";
+
+
 
 export default function Dashboard() {
-  const [user, setUser] = useRecoilState(userAtom);
-  const [isSidebarOpen, setSidebarOpen] = useState(false);
+  const user = useRecoilValue(userAtom);
 
-  const handleLogout = () => {
-    Cookies.remove("user");
-    setUser(null);
-    console.log("User logged out");
-  };
-
-  const toggleSidebar = () => {
-    setSidebarOpen(!isSidebarOpen);
-  };
 
   return (
-    <div className="flex h-screen bg-gray-100 overflow-hidden">
-      {/* Sidebar */}
-      <AnimatePresence>
-        <Sidebar isSidebarOpen={isSidebarOpen} toggleSidebar={toggleSidebar} isHome={true}/>
-      </AnimatePresence>
-
-      
-
-      {/* Main Content */}
+  
       <div className="flex-1 flex flex-col overflow-hidden">
-        {/* Header */}
-        <header className="bg-white shadow-sm">
-          <div className="max-w-7xl mx-auto py-4 px-4 sm:px-6 lg:px-8 flex justify-between items-center">
-            <div className="flex items-center">
-              <button
-                onClick={toggleSidebar}
-                className="md:hidden mr-4 text-gray-600 hover:text-gray-900"
-              >
-                <FaBars size={24} />
-              </button>
-              <h2 className="text-2xl font-semibold text-gray-800">
-                Dashboard
-              </h2>
-            </div>
-            <div className="flex items-center space-x-4">
-              <div className="relative group">
-                <div className="flex items-center space-x-2 cursor-pointer">
-                  <img
-                    src="https://th.bing.com/th/id/OIP.n0waXJvNzJqj3wmDBfS1ZwHaHa?w=165&h=180&c=7&r=0&o=5&dpr=2&pid=1.7"
-                    alt="User Avatar"
-                    className="rounded-full w-10 h-10"
-                  />
-                  <span className="text-gray-800 font-medium">
-                    {user?.name || "User"}
-                  </span>
-                </div>
-                <div className="absolute right-0 w-48 mt-2 py-2 bg-white rounded-md shadow-xl z-20 hidden group-hover:block">
-                  <button
-                    onClick={handleLogout}
-                    className="block px-4 py-2 text-sm text-gray-700 hover:bg-indigo-500 hover:text-white w-full text-left"
-                  >
-                    <FaSignOutAlt className="inline-block mr-2" /> Logout
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </header>
+        
 
         {/* Dashboard Content */}
         <main className="flex-1 overflow-x-hidden overflow-y-auto bg-gray-100">
@@ -117,7 +42,6 @@ export default function Dashboard() {
           </div>
         </main>
       </div>
-    </div>
   );
 }
