+++ Admin_Teacher_Frontend/src/pages/Manage_Images/ManageImages.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ManageImages = () => {
+  return (
+    <div>
+      manage image page
+    </div>
+  )
+}
+
+export default ManageImages
