+++ Admin_Teacher_Frontend/src/pages/SubjectStdPage/SubjectStdPage.tsx
@@ -5,10 +5,7 @@ import { useRecoilState } from 'recoil'
 import { motion, AnimatePresence } from 'framer-motion'
 import { stdSubAtom } from '../../state/stdSubAtom'
 import { teachersAtom } from '../../state/teachersAtom'
-import { userAtom } from "../../state/userAtom"
-import { FaPlus, FaTrash, FaPencilAlt, FaSave, FaTimes, FaGraduationCap, FaChalkboardTeacher, FaBars, FaSignOutAlt } from 'react-icons/fa'
-import Cookies from "js-cookie"
-import Sidebar from '../Dashbord/Sidebar'
+import { FaPlus, FaTrash, FaPencilAlt, FaSave, FaTimes, FaGraduationCap, FaChalkboardTeacher } from 'react-icons/fa'
 import axios from 'axios'
 import { BACKEND_URL } from '../../config'
 import { toast } from 'react-toastify'
@@ -18,7 +15,6 @@ import ConformDeletePopUp from '../../components/conformation/ConformDeletePopUp
 export default function SubjectStandardPage() {
   const [stdSubState, setStdSubState] = useRecoilState<any>(stdSubAtom)
   const [teachers, setTeachers] = useRecoilState(teachersAtom)
-  const [user, setUser] = useRecoilState(userAtom)
   const [isLoading , setIsLoading] = useState(false)
 
   const [addStdText, setAddStdText] = useState('')
@@ -28,7 +24,6 @@ export default function SubjectStandardPage() {
   const [editStdName, setEditStdName] = useState('')
   const [editStdId, setEditStdId] = useState('')
 
-  const [isSidebarOpen, setSidebarOpen] = useState(false)
   const [isConformDelete, setIsConformDelete] = useState(false)
   const [deleteData, setDeleteData] = useState<any>({})
 
@@ -170,18 +165,10 @@ export default function SubjectStandardPage() {
     }
   }
 
-  const handleLogout = () => {
-    Cookies.remove("user")
-    setUser(null)
-    console.log("User logged out")
-  }
-
-  const toggleSidebar = () => {
-    setSidebarOpen(!isSidebarOpen)
-  }
+ 
 
   return (
-    <div className="flex h-screen bg-gray-100 overflow-hidden">
+    <div className="">
       {/* Sidebar */}
       {
         isLoading && <Loading/>
@@ -190,41 +177,12 @@ export default function SubjectStandardPage() {
         isConformDelete && <ConformDeletePopUp show={isConformDelete} handleClose={() => setIsConformDelete(false)} handleConfirm={() => handleDeleteTeacherFromSubject(deleteData.standardId, deleteData.subjectName, deleteData.teacherId)} />
       }
       
-     <Sidebar isSidebarOpen={isSidebarOpen} toggleSidebar={toggleSidebar}/>
+   
 
       {/* Main Content */}
       <div className="flex-1 flex flex-col overflow-hidden">
         {/* Header */}
-        <header className="bg-white shadow-sm">
-          <div className="max-w-7xl mx-auto py-4 px-4 sm:px-6 lg:px-8 flex justify-between items-center">
-            <div className="flex items-center">
-              <button onClick={toggleSidebar} className="md:hidden mr-4 text-gray-600 hover:text-gray-900">
-                <FaBars size={24} />
-              </button>
-              <h2 className="text-2xl font-semibold text-gray-800">Manage Standards and Subjects</h2>
-            </div>
-            <div className="flex items-center space-x-4">
-              <div className="relative group">
-                <div className="flex items-center space-x-2 cursor-pointer">
-                  <img
-                    src="https://th.bing.com/th/id/OIP.n0waXJvNzJqj3wmDBfS1ZwHaHa?w=165&h=180&c=7&r=0&o=5&dpr=2&pid=1.7"
-                    alt="User Avatar"
-                    className="rounded-full w-10 h-10"
-                  />
-                  <span className="text-gray-800 font-medium">{user?.name || "User"}</span>
-                </div>
-                <div className="absolute right-0 w-48 mt-2 py-2 bg-white rounded-md shadow-xl z-20 hidden group-hover:block">
-                  <button
-                    onClick={handleLogout}
-                    className="block px-4 py-2 text-sm text-gray-700 hover:bg-indigo-500 hover:text-white w-full text-left"
-                  >
-                    <FaSignOutAlt className="inline-block mr-2" /> Logout
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </header>
+     
 
         {/* Page Content */}
         <main className="flex-1 overflow-x-hidden overflow-y-auto bg-gray-100">
